refactor: move cell generation handling to demo.utils
