+++ utils/container.utils.ts
@@ -1,64 +1,4 @@
-import { Container, ContainerSize, Cell, Item, Input } from '../interfaces/inventory'
-import { generateId } from '@/utils/id.utils'
-import { generateArray, findLastIndex } from '@/utils/array.utils'
-
-/**
- * Return formatted container
- */
-type createContainerOptions = {
-  id?: number
-  name?: string
-  size?: ContainerSize
-  items?: Input[]
-}
-
-export const createContainer = (options: createContainerOptions): Container => {
-  const id = options.id ?? generateId()
-  const name = options.name ?? ''
-  const size = options.size ?? ([8, 2] as ContainerSize)
-  const cells = generateCells(id, size, options.items || [])
-
-  return { ...options, id, name, cells, size }
-}
-
-/**
- * Generate container cells from input data
- */
-export const generateCells = (id: number, size: ContainerSize, currentCells: Input[]): Cell[] => {
-  const calculateTotalCells = (containerSize: ContainerSize, currentCells: Input[]) => {
-    const containerCols = containerSize[0]
-    const containerRows = containerSize[1]
-
-    const lastFilledIndex = findLastIndex(currentCells, (cell) => cell.item !== null)
-    const lastFilledRow = lastFilledIndex === -1 ? 0 : Math.floor(lastFilledIndex / containerSize[0])
-
-    const currentNeededCells = (lastFilledRow + 1) * containerCols
-
-    const minimumCells = containerCols * containerRows
-
-    if (currentNeededCells < minimumCells) {
-      return minimumCells
-    }
-
-    const currentFilledCellsAmount = currentCells.filter((cell) => cell.item !== null).length
-
-    if (currentFilledCellsAmount >= currentNeededCells - 1) {
-      return currentNeededCells + containerCols
-    }
-
-    return currentNeededCells
-  }
-
-  const totalCells = calculateTotalCells(size, currentCells)
-
-  const cells = generateArray(totalCells).map((_, cellId) => ({
-    id: cellId,
-    path: [id, cellId],
-    item: currentCells?.[cellId]?.item || null,
-  }))
-
-  return cells
-}
+import { Container, Cell, Item } from '../interfaces/inventory'
 
 /**
  * Find item in cell
@@ -128,7 +68,6 @@ export const depositFirstAvailableCell = (cells: Cell[], item: Item): Cell[] => 
 }
 
 export default {
-  createContainer,
   findCell,
   clearCell,
   setCell,
